Create form & server action to password reset
Create a new form that will display when the user clicks the
link inside the password-reset-email sent to them. This form
will allow them to add a new password.

Create a server action that will do all the validation at the
server-side and update the user password if all validation
succeeds.